+++ docs/changelog.rst
@@ -47,6 +47,7 @@ Changes
   * Existing recipes are updated automatically on import.
 
 * Figure properties (see :class:`aspecd.plotting.FigureProperties`) are applied to the figure *before* the actual plotting is done, as some plotters need to know the (final) figure size or else.
+* Axes properties (see :class:`aspecd.plotting.AxesProperties`) are applied to the axes *before* the actual plotting is done, as some plotters need to know the (final) axes limits or else.
 
 
 Fixes